requirements.txt
lida
lida[transformers]
torch
streamlit
python-dotenv
openai
pillow
pandas
time